properly merged files, by editing new into old
