--- src/client/features/frontend/allStudents.jsx
@@ -1,18 +1,0 @@
-// AllStudents.js
-
-import React from 'react';
-
-function AllStudents() {
-  // Fetch and display the list of students
-  return (
-    <div>
-      <h2>All Students</h2>
-      {/* Your list of students goes here */}
-    </div>
-  );
-}
-
-export default AllStudents;
-
-
-//this is teporary and will be fixed when mereged, for now it is in place so that the sever works.
