--- src/client/features/frontend/studentList.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const StudentList = ({ students }) => {
-  return (
-    <div>
-      <h2>Student List</h2>
-      <ul>
-        {students.map((student, index) => (
-          <li key={index}>{student}</li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-export default StudentList;
