--- src/client/store/app.js
@@ -1,21 +1,0 @@
-// App.js
-
-import React from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import AllStudents from "./AllStudents"; // Import your AllStudents component
-import Navbar from "../layout/Navbar";
-
-function App() {
-  return (
-    <Router>
-      <Switch>
-        <Route path="/students">
-          <AllStudents />
-        </Route>
-        {/* Other routes can be defined here */}
-      </Switch>
-    </Router>
-  );
-}
-
-export default App;
